Rename file to match our code style recommendation
https://docs.sourcegraph.com/dev/background-information/web/web_app#naming-files
