+++ client/jetbrains/webview/src/search/types.d.ts
@@ -1,6 +1,6 @@
 import type { SearchPatternType } from '@sourcegraph/search'
 
-import type { Request } from './jsToJavaBridgeUtil'
+import type { Request } from './js-to-java-bridge'
 
 /* Add global functions to global window object */
 declare global {
